INTERLOK-639 #comment Updated MleMarker#sequenceNumber to be a long. INTERLOK-639 #time 20m
Updated MleMarker#sequenceNumber to be a long. AdaptrisMessageImp
nextSequenceNumber changed to reflect the change. No tests appear
broken but cirrus will be broken with an incompatible type on compile
when the nightly is released.

+++ adapter/src/main/java/com/adaptris/core/AdaptrisMessageImp.java
@@ -322,13 +322,13 @@ public abstract class AdaptrisMessageImp implements AdaptrisMessage, Cloneable {
   }
 
   private MleMarker getNextMleMarker(MessageEventGenerator meg, boolean successful, String confId) {
-    int seq = nextSequenceNumber();
+    long seq = nextSequenceNumber();
     this.addMetadata(CoreConstants.MLE_SEQUENCE_KEY, String.valueOf(seq));
     MleMarker mleMarker = new MleMarker(meg, successful, seq, guidGenerator.create(meg), confId);
     return mleMarker;
   }
 
-  private int nextSequenceNumber() {
+  private long nextSequenceNumber() {
     int result= 0;
     if (containsKey(CoreConstants.MLE_SEQUENCE_KEY)) {
       try {
